chore: Move common middleware to the common folder
