+++ app/auth/middleware.js
@@ -33,6 +33,7 @@ function processAuthResponse() {
           locations,
           roles: decodedAccessToken.authorities,
         })
+
         // copy any previous session properties ignoring grant or any that already exist
         Object.keys(previousSession).forEach(key => {
           if (req.session[key]) {
@@ -43,6 +44,7 @@ function processAuthResponse() {
           }
           req.session[key] = previousSession[key]
         })
+
         next()
       })
     } catch (error) {
