+++ common/middleware/index.js
@@ -1,0 +1,7 @@
+const setDatePeriod = require('./set-date-period')
+const setDateRange = require('./set-date-range')
+
+module.exports = {
+  setDatePeriod,
+  setDateRange,
+}
