+++ common/middleware/set-date-period.js
@@ -1,0 +1,6 @@
+function setDatePeriod(req, res, next, period) {
+  res.locals.period = period
+  next()
+}
+
+module.exports = setDatePeriod
